fix: shifted a hook and changed the query key for some hooks
