+++ src/slices/hatch/admin/hooks/bookingHooks.ts
@@ -1,5 +1,5 @@
 import { TBooking } from '@slices/hatch/booking-page/types';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { WithId } from 'mongodb';
 
 import { TBatchBookingResponse } from '@/app/api/types';
@@ -7,6 +7,27 @@ import {
   fetchBatchAddBooking,
   fetchBatchDeleteBooking,
 } from '@/slices/hatch/admin/apiCalls/bookingApiCalls';
+import { fetchAllBookings } from '@/slices/hatch/booking-page/apiCalls/bookingApiCalls';
+import { formatDateForKey } from '@/slices/hatch/booking-page/utils';
+
+export const useFetchAllBookingsHook = (
+  startDate: Date,
+  endDate: Date,
+  enabled: boolean,
+) => {
+  return useQuery<TBooking[], Error>({
+    queryKey: [
+      'allBookings',
+      formatDateForKey(startDate),
+      formatDateForKey(endDate),
+    ],
+    queryFn: () => fetchAllBookings(startDate, endDate),
+    staleTime: 1000 * 60 * 5, // 5 minutes
+    refetchOnWindowFocus: false,
+    refetchOnMount: false,
+    enabled: enabled,
+  });
+};
 
 export const useBatchAddRoomBookingHook = () => {
   const queryClient = useQueryClient();
@@ -15,7 +36,7 @@ export const useBatchAddRoomBookingHook = () => {
     mutationFn: fetchBatchAddBooking,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['roomAvailabilities'] });
-      queryClient.invalidateQueries({ queryKey: ['userBookings'] });
+      queryClient.invalidateQueries({ queryKey: ['allBookings'] });
     },
   });
 };
@@ -27,7 +48,7 @@ export const useBatchDeleteRoomBookingHook = () => {
     mutationFn: fetchBatchDeleteBooking,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['roomAvailabilities'] });
-      queryClient.invalidateQueries({ queryKey: ['userBookings'] });
+      queryClient.invalidateQueries({ queryKey: ['allBookings'] });
     },
   });
 };
