+++ src/slices/hatch/booking-page/hooks/bookingHooks.ts
@@ -1,18 +1,14 @@
 import { TBooking } from '@slices/hatch/booking-page/types';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { startOfDay } from 'date-fns';
 
 import {
   fetchAddBooking,
-  fetchAllBookings,
   fetchAvailabilities,
   fetchDeleteBooking,
   fetchUserBookings,
 } from '@/slices/hatch/booking-page/apiCalls/bookingApiCalls';
 import { RoomAvailabilities } from '@/slices/hatch/booking-page/components/TimePicker';
-
-const formatDateForKey = (date: Date) =>
-  startOfDay(date).toISOString().split('T')[0];
+import { formatDateForKey } from '@/slices/hatch/booking-page/utils';
 
 export const useAddRoomBookingHook = () => {
   const queryClient = useQueryClient();
@@ -60,24 +56,5 @@ export const useDeleteBookingHook = () => {
       queryClient.invalidateQueries({ queryKey: ['userBookings'] });
       queryClient.invalidateQueries({ queryKey: ['roomAvailabilities'] });
     },
-  });
-};
-
-export const useFetchAllBookingsHook = (
-  startDate: Date,
-  endDate: Date,
-  enabled: boolean,
-) => {
-  return useQuery<TBooking[], Error>({
-    queryKey: [
-      'allBookings',
-      formatDateForKey(startDate),
-      formatDateForKey(endDate),
-    ],
-    queryFn: () => fetchAllBookings(startDate, endDate),
-    staleTime: 1000 * 60 * 5, // 5 minutes
-    refetchOnWindowFocus: false,
-    refetchOnMount: false,
-    enabled: enabled,
   });
 };
