+++ src/slices/hatch/booking-page/utils.ts
@@ -1,5 +1,8 @@
-import { addMinutes } from 'date-fns';
+import { addMinutes, startOfDay } from 'date-fns';
 
 export function add30Minutes(date: Date): Date {
   return addMinutes(date, 30);
 }
+
+export const formatDateForKey = (date: Date) =>
+  startOfDay(date).toISOString().split('T')[0];
